Move new hooks into components/hooks directory
