+++ packages/lesswrong/components/hooks/useCrosspostApolloClient.tsx
@@ -1,6 +1,6 @@
 import type { ApolloClient, NormalizedCacheObject } from '@apollo/client';
-import { createApolloClient } from "../../../client/apolloClient";
-import { fmCrosspostBaseUrlSetting } from "../../../lib/instanceSettings";
+import { createApolloClient } from "../../client/apolloClient";
+import { fmCrosspostBaseUrlSetting } from "../../lib/instanceSettings";
 
 let apolloClient: ApolloClient<NormalizedCacheObject> | undefined;
